OPCUAClient修复当变量为string类型时的数值过滤失败导致添加订阅失效

+++ src/Plugins/ThingsGateway.Foundation.Adapter.OPCUA/OPCUAClient/OPCNode.cs
@@ -16,7 +16,7 @@ public class OPCNode
     public string OPCURL { get; set; } = "opc.tcp://127.0.0.1:49320";
     public int UpdateRate { get; set; } = 1000;
     public int GroupSize { get; set; } = 500;
-    public float DeadBand { get; set; } = 0;
+    public double DeadBand { get; set; } = 0;
     public int ReconnectPeriod { get; set; } = 5000;
     public bool IsUseSecurity { get; set; } = true;
